requirements.txt
pyyaml
fastapi
uvicorn
kubernetes
click
openai
python-multipart
